Don't use Files.createTempFile() while loading libjssl.so
During initialization of OpenSSLFIPSProvider, we load libjssl.so
shipped as a resource from the jar file built by maven. We create
a temporary file, copy the resource into it and call
System.loadLibrary() on the temporary file.

The Files.createTempFile() method needs a SecureRandom. If
OpenSSLFIPSProvider is configured as the provider with top
priority, this will attempt to initialize OpenSSLDRBG causing
a recursive attempt to load libjssl.so and failing with an NPE.

This commit removes the use of Files.createTempFile() to make sure
no other security function is not invoked while loading libjssl.so
during initialization.